fix api.

+++ module/src/display/block.rs
@@ -8,6 +8,6 @@ pub struct DisplayBlock {
     pub time: i64,
     pub app_hash: String,
     pub proposer: String,
-    pub tx_count: usize,
+    pub tx_count: i64,
     pub size: i64,
 }
